refactor(ariel): harden interface JS to the strict type/lint standard
Bring the ariel logbook interface under the enforced front-end gates,
matching the tuning and channel_finder interfaces.

- Remove every per-file @ts-nocheck header and type-clean each JS module
  under strict checkJs: JSDoc casts and null-guards, an ApiError class
  carrying status/code, and a shared messageOf helper for catch reads.
- Split entries.js into focused modules under the max-lines cap:
  entries.js (lifecycle/list), entries-detail.js (modal detail view),
  entries-form.js (create/draft form), entries-helpers.js (pure helpers).
  The public export surface is preserved so app.js wiring is untouched;
  the module-level state each cluster owns moves with it.
- Drop the two ariel eslint exemptions (no-unused-vars, max-lines) and
  remove two dead parameters flagged under strict checking.
- Dedupe repeated error-state, attachment-preview, and JSON-request
  construction into shared helpers.

Behavior-preserving throughout; typecheck, lint, and test:js stay green.
