+++ src/osprey/interfaces/ariel/static/js/api.js
@@ -1,5 +1,4 @@
-// @ts-nocheck
-// TODO(frontend-hardening Pn): remove & fix types when this interface is retrofitted (P2–P5)
+// @ts-check
 /**
  * ARIEL API Client
  *
@@ -9,20 +8,60 @@
 const API_BASE = '/api';
 
 /**
- * Build an Error from a non-OK response.
+ * Structured API error carrying the HTTP status code and an optional
+ * machine-readable `code` discriminator from the response body.
+ */
+export class ApiError extends Error {
+  /**
+   * @param {number} status - HTTP status code.
+   * @param {string} message - Error detail message.
+   * @param {string} [code] - Machine-readable discriminator (e.g. "auth_required").
+   */
+  constructor(status, message, code) {
+    super(message);
+    /** @type {number} */
+    this.status = status;
+    /** @type {string|undefined} */
+    this.code = code;
+    this.name = 'ApiError';
+  }
+}
+
+/**
+ * Build an ApiError from a non-OK response.
  *
  * Preserves the HTTP `status` and any machine-readable `code` discriminator from
  * the JSON body (e.g. "auth_required") so callers can branch on them — for
  * example, to prompt for logbook credentials instead of showing a generic error.
  * @param {Response} response - The failed fetch response
- * @returns {Promise<Error>} An Error with `.status` and optional `.code`
+ * @returns {Promise<ApiError>} An ApiError with `.status` and optional `.code`
  */
 async function errorFromResponse(response) {
   const body = await response.json().catch(() => ({}));
-  const error = new Error(body.detail || `HTTP ${response.status}`);
-  error.status = response.status;
-  if (body.code) error.code = body.code;
-  return error;
+  return new ApiError(response.status, body.detail || `HTTP ${response.status}`, body.code);
+}
+
+/**
+ * Issue a JSON-body request, shared by `post` and `put` (which only differ
+ * in HTTP method).
+ * @param {string} method - HTTP method ('POST', 'PUT')
+ * @param {string} endpoint - API endpoint
+ * @param {any} data - Request body
+ * @returns {Promise<any>} Response data
+ */
+async function jsonRequest(method, endpoint, data) {
+  const response = await fetch(API_BASE + endpoint, {
+    method,
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(data),
+  });
+
+  if (!response.ok) {
+    throw await errorFromResponse(response);
+  }
+  return response.json();
 }
 
 /**
@@ -32,8 +71,8 @@ export const api = {
   /**
    * Make a GET request.
    * @param {string} endpoint - API endpoint
-   * @param {Object} params - Query parameters
-   * @returns {Promise<Object>} Response data
+   * @param {any} params - Query parameters
+   * @returns {Promise<any>} Response data
    */
   async get(endpoint, params = {}) {
     const url = new URL(API_BASE + endpoint, window.location.origin);
@@ -53,43 +92,21 @@ export const api = {
   /**
    * Make a POST request.
    * @param {string} endpoint - API endpoint
-   * @param {Object} data - Request body
-   * @returns {Promise<Object>} Response data
+   * @param {any} data - Request body
+   * @returns {Promise<any>} Response data
    */
   async post(endpoint, data = {}) {
-    const response = await fetch(API_BASE + endpoint, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (!response.ok) {
-      throw await errorFromResponse(response);
-    }
-    return response.json();
+    return jsonRequest('POST', endpoint, data);
   },
 
   /**
    * Make a PUT request.
    * @param {string} endpoint - API endpoint
-   * @param {Object} data - Request body
-   * @returns {Promise<Object>} Response data
+   * @param {any} data - Request body
+   * @returns {Promise<any>} Response data
    */
   async put(endpoint, data = {}) {
-    const response = await fetch(API_BASE + endpoint, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (!response.ok) {
-      throw await errorFromResponse(response);
-    }
-    return response.json();
+    return jsonRequest('PUT', endpoint, data);
   },
 };
 
@@ -99,7 +116,7 @@ export const api = {
 export const capabilitiesApi = {
   /**
    * Fetch available search modes and their tunable parameters.
-   * @returns {Promise<Object>} Capabilities response
+   * @returns {Promise<any>} Capabilities response
    */
   async get() {
     return api.get('/capabilities');
@@ -112,8 +129,8 @@ export const capabilitiesApi = {
 export const searchApi = {
   /**
    * Execute a search query.
-   * @param {Object} params - Search parameters
-   * @returns {Promise<Object>} Search results
+   * @param {any} params - Search parameters
+   * @returns {Promise<any>} Search results
    */
   async search(params) {
     const request = {
@@ -142,8 +159,8 @@ export const entriesApi = {
 
   /**
    * List entries with pagination.
-   * @param {Object} params - List parameters
-   * @returns {Promise<Object>} Paginated entries
+   * @param {any} params - List parameters
+   * @returns {Promise<any>} Paginated entries
    */
   async list(params = {}) {
     return api.get('/entries', {
@@ -160,7 +177,7 @@ export const entriesApi = {
   /**
    * Get a single entry by ID.
    * @param {string} entryId - Entry ID
-   * @returns {Promise<Object>} Entry data
+   * @returns {Promise<any>} Entry data
    */
   async get(entryId) {
     return api.get(`/entries/${entryId}`);
@@ -168,8 +185,8 @@ export const entriesApi = {
 
   /**
    * Create a new entry.
-   * @param {Object} data - Entry data
-   * @returns {Promise<Object>} Created entry
+   * @param {any} data - Entry data
+   * @returns {Promise<any>} Created entry
    */
   async create(data) {
     return api.post('/entries', {
@@ -187,9 +204,9 @@ export const entriesApi = {
 
   /**
    * Create a new entry with file attachments via multipart form.
-   * @param {Object} data - Entry data fields
+   * @param {any} data - Entry data fields
    * @param {FileList|File[]} files - Files to attach
-   * @returns {Promise<Object>} Created entry with attachment_count
+   * @returns {Promise<any>} Created entry with attachment_count
    */
   async createWithAttachments(data, files) {
     const formData = new FormData();
@@ -228,7 +245,7 @@ export const draftsApi = {
   /**
    * Get a draft by ID.
    * @param {string} draftId - Draft ID
-   * @returns {Promise<Object>} Draft data
+   * @returns {Promise<any>} Draft data
    */
   async get(draftId) {
     return api.get(`/drafts/${draftId}`);
@@ -241,7 +258,7 @@ export const draftsApi = {
 export const statusApi = {
   /**
    * Get service status.
-   * @returns {Promise<Object>} Status information
+   * @returns {Promise<any>} Status information
    */
   async get() {
     return api.get('/status');
@@ -255,6 +272,9 @@ export const configApi = {
   async get() {
     return api.get('/config');
   },
+  /**
+   * @param {string} content - Raw config.yml text.
+   */
   async update(content) {
     return api.put('/config', { content });
   },
