+++ src/osprey/interfaces/ariel/static/js/search.js
@@ -1,5 +1,4 @@
-// @ts-nocheck
-// TODO(frontend-hardening Pn): remove & fix types when this interface is retrofitted (P2–P5)
+// @ts-check
 /**
  * ARIEL Search Module
  *
@@ -13,20 +12,33 @@ import {
   renderDiagnosticsBar,
   renderLoading,
   renderEmptyState,
+  renderErrorState,
   escapeHtml,
 } from './components.js';
 import { getCurrentMode, getAdvancedParams, closeAdvancedPanel } from './advanced-options.js';
 
+/**
+ * @typedef {Object} SearchResults
+ * @property {string} [answer]
+ * @property {string[]} [sources]
+ * @property {string[]} [search_modes_used]
+ * @property {number} [execution_time_ms]
+ * @property {number} total_results
+ * @property {import('./components.js').Entry[]} [entries]
+ * @property {import('./components.js').Diagnostic[]} [diagnostics]
+ */
+
 // Search state
 let currentQuery = '';
 let isSearching = false;
+/** @type {SearchResults|null} */
 let lastResults = null;
 
 /**
  * Initialize search module.
  */
 export function initSearch() {
-  const searchInput = document.getElementById('search-input');
+  const searchInput = /** @type {HTMLInputElement|null} */ (document.getElementById('search-input'));
   const searchBtn = document.getElementById('search-btn');
 
   // Search input enter key
@@ -52,7 +64,7 @@ export function initSearch() {
       searchInput?.focus();
     }
     // Escape to clear search
-    if (e.key === 'Escape' && document.activeElement === searchInput) {
+    if (e.key === 'Escape' && document.activeElement === searchInput && searchInput) {
       searchInput.value = '';
       searchInput.blur();
     }
@@ -61,10 +73,10 @@ export function initSearch() {
 
 /**
  * Perform a search.
- * @param {string} query - Optional query override
+ * @param {string|null} [query] - Optional query override
  */
 export async function performSearch(query = null) {
-  const searchInput = document.getElementById('search-input');
+  const searchInput = /** @type {HTMLInputElement|null} */ (document.getElementById('search-input'));
   const resultsContainer = document.getElementById('search-results');
 
   query = query || searchInput?.value?.trim();
@@ -83,6 +95,7 @@ export async function performSearch(query = null) {
 
   // Get mode and advanced params from the unified capabilities-driven UI
   const mode = getCurrentMode();
+  /** @type {Object<string, *>} */
   const advancedParams = getAdvancedParams();
 
   try {
@@ -98,12 +111,7 @@ export async function performSearch(query = null) {
   } catch (error) {
     console.error('Search failed:', error);
     if (resultsContainer) {
-      resultsContainer.innerHTML = `
-        <div class="empty-state">
-          <h3 class="empty-state-title text-error">Search Failed</h3>
-          <p class="empty-state-text">${escapeHtml(error.message)}</p>
-        </div>
-      `;
+      resultsContainer.innerHTML = renderErrorState('Search Failed', error);
     }
   } finally {
     isSearching = false;
@@ -112,7 +120,7 @@ export async function performSearch(query = null) {
 
 /**
  * Render search results.
- * @param {Object} results - Search results from API
+ * @param {SearchResults} results - Search results from API
  * @param {string} mode - The search mode selected by the user (e.g. 'keyword', 'semantic')
  */
 function renderSearchResults(results, mode = 'keyword') {
@@ -132,8 +140,8 @@ function renderSearchResults(results, mode = 'keyword') {
   }
 
   // Diagnostics bar if issues detected
-  if (results.diagnostics?.length > 0) {
-    html += renderDiagnosticsBar(results.diagnostics);
+  if ((results.diagnostics?.length ?? 0) > 0) {
+    html += renderDiagnosticsBar(results.diagnostics ?? []);
   }
 
   // Results header
@@ -150,10 +158,10 @@ function renderSearchResults(results, mode = 'keyword') {
   `;
 
   // Results list
-  if (results.entries?.length > 0) {
+  if ((results.entries?.length ?? 0) > 0) {
     const sourcesSet = results.sources?.length ? new Set(results.sources) : null;
     html += '<div class="results-list">';
-    results.entries.forEach(entry => {
+    (results.entries ?? []).forEach(entry => {
       const isCited = sourcesSet ? sourcesSet.has(entry.entry_id) : false;
       html += renderEntryCard(entry, isCited);
     });
@@ -172,7 +180,7 @@ function renderSearchResults(results, mode = 'keyword') {
  * Clear search results.
  */
 export function clearSearch() {
-  const searchInput = document.getElementById('search-input');
+  const searchInput = /** @type {HTMLInputElement|null} */ (document.getElementById('search-input'));
   const resultsContainer = document.getElementById('search-results');
 
   if (searchInput) searchInput.value = '';
